# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Arabhardware API Documentation",
        "description": "This is the API documentation for the Arabhardware project. It provides a comprehensive guide to interacting with the API endpoints.",
        "contact": {
            "name": "Support Team",
            "email": "sbilal@arabhardware.net"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Arabhardware API Server"
        }
    ],
    "paths": {
...
        "/api/v1/ahwstore/sales-orders": {
            "get": {
                "tags": [
                    "SalesOrders"
                ],
                "summary": "Get all sales orders",
                "operationId": "5ba2c59858f232be8837f8ef6b671144",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SalesOrderResponse"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            },
            "post": {
                "tags": [
                    "SalesOrders"
                ],
                "summary": "Create a new sales order",
                "operationId": "1e203b8aa6ba4a7d7a293568c505ecb4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SalesOrderRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sales order created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SalesOrderResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/v1/ahwstore/sales-orders/{id}": {
            "get": {
                "tags": [
                    "SalesOrders"
                ],
                "summary": "Get a specific sales order",
                "operationId": "9551eb82a03b2576caf70cb41c2782f0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Sales Order ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SalesOrderResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Sales order not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "SalesOrders"
                ],
                "summary": "Update a sales order",
                "operationId": "ecd869dfac97234196cc2431587e8470",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Sales Order ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SalesOrderUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sales order updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SalesOrderResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "404": {
                        "description": "Sales order not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "SalesOrders"
                ],
                "summary": "Delete a sales order",
                "operationId": "2a80deb8bc938f69b17f38647cd2bced",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Sales Order ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Sales order deleted successfully"
                    },
                    "404": {
                        "description": "Sales order not found"
                    }
                }
            }
        },
        "/api/v1/ahwstore/sales-orders/{id}/status/confirmed": {
            "post": {
                "tags": [
                    "SalesOrders"
                ],
                "summary": "Mark a sales order as Confirmed",
                "operationId": "50f6d8d2f9ece64a9882fb0931a7b13f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Sales Order ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sales order status changed to Confirmed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 0
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Sales order status has been changed to 'Confirmed'."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/ahwstore/sales-orders/{id}/status/void": {
            "post": {
                "tags": [
                    "SalesOrders"
                ],
                "summary": "Mark a sales order as Void",
                "operationId": "3d75bd735b95ae09ea4857eb151640c3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Sales Order ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sales order status changed to Void",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 0
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "The status of the Sales Order has been changed to void."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/ahwstore/sales-orders/status/confirmed": {
            "post": {
                "tags": [
                    "SalesOrders"
                ],
                "summary": "Bulk confirm sales orders",
                "operationId": "8023f0926738082584395fce75bf6442",
                "parameters": [
                    {
                        "name": "salesorder_ids",
                        "in": "query",
                        "description": "Comma-separated list of sales order IDs",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "4815000000045208,4815000000045274,4815000000045340"
                        }
                    },
                    {
                        "name": "organization_id",
                        "in": "query",
                        "description": "Organization ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "10234695"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bulk sales order status changed to Confirmed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 0
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Sales order status has been changed to Confirmed."
                                        },
                                        "data": {
                                            "properties": {
                                                "email_success_info": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "code": {
                                                                "type": "integer",
                                                                "example": 0
                                                            },
                                                            "message": {
                                                                "type": "string",
                                                                "example": "The status of the sales order has been changed to Open."
                                                            },
                                                            "ids": {
                                                                "type": "array",
                                                                "items": {
                                                                    "type": "string",
                                                                    "example": "4815000000045208,4815000000045274,4815000000045340"
                                                                }
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "email_error_info": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "SalesOrderResponse": {
                "properties": {
                    "salesorder_id": {
                        "type": "string",
                        "example": "4815000000045208"
                    },
                    "customer_id": {
                        "type": "integer",
                        "example": 12345
                    },
                    "customer_name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-01-15"
                    },
                    "shipment_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-01-20"
                    },
                    "delivery_method": {
                        "type": "string",
                        "example": "Standard"
                    },
                    "reference_number": {
                        "type": "string",
                        "example": "REF-001"
                    },
                    "notes": {
                        "type": "string",
                        "example": "Order notes"
                    },
                    "terms": {
                        "type": "string",
                        "example": "Net 30"
                    },
                    "total": {
                        "type": "number",
                        "format": "float",
                        "example": 1250.5
                    },
                    "status": {
                        "type": "string",
                        "example": "confirmed"
                    },
                    "line_items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/LineItem"
                        }
                    }
                },
                "type": "object"
            },
            "SalesOrderRequest": {
                "required": [
                    "customer_id",
                    "date",
                    "line_items"
                ],
                "properties": {
                    "customer_id": {
                        "type": "integer",
                        "example": 12345
                    },
                    "date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-01-15"
                    },
                    "shipment_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-01-20"
                    },
                    "delivery_method": {
                        "type": "string",
                        "example": "Standard"
                    },
                    "reference_number": {
                        "type": "string",
                        "example": "REF-001"
                    },
                    "notes": {
                        "type": "string",
                        "example": "Order notes"
                    },
                    "terms": {
                        "type": "string",
                        "example": "Net 30"
                    },
                    "line_items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/LineItemRequest"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}